Create cloud_objects crate
