+++ crates/cloud_objects/src/auth/mod.rs
@@ -1,0 +1,3 @@
+pub mod user_uid;
+
+pub use user_uid::{TEST_USER_EMAIL, TEST_USER_UID, UserUid};
