+++ crates/cloud_objects/src/cloud_object/generic_cloud_object.rs
