+++ crates/cloud_objects/src/cloud_object/generic_string_model.rs
