+++ crates/cloud_objects/src/cloud_object/models/mod.rs
