+++ crates/cloud_objects/src/cloud_object/server_object.rs
