+++ crates/cloud_objects/src/drive/sharing.rs
