+++ crates/warp_server_client/src/lib.rs
@@ -6,3 +6,4 @@ pub mod ids;
 pub mod persistence;
 
 pub use auth::UserUid;
+pub use cloud_objects::server_id_traits;
